Ensure name does not refer to specific quantity
Although we're specifically asking the LLM not to include any references
to a specific quantity and produce a generic product name, some models
may not be strong enough to follow the instruction. gemma3:4b certainly
seems to ignore it often.

Query the model with a much more specific prompt to maximize the
probability of getting a generic name back.

Signed-off-by: Vangelis Koukis <vkoukis@gmail.com>